fix: move lock_order to lib layer, fix mypy, merge develop
- Move lock_order.py from nexus.core to nexus.lib (architecture boundary:
  lib must not import core, but core and raft can import lib)
- Fix mypy: replace threading.local() with typed dict keyed by thread id
- Update all imports and doc references
- Merge latest develop
